astype can not be generic

+++ spinn_front_end_common/utility_models/reverse_ip_tag_multicast_source_machine_vertex.py
@@ -252,7 +252,7 @@ class ReverseIPTagMulticastSourceMachineVertex(
         """
         if is_array_list(send_buffer_times):
             counts = numpy.bincount(
-                numpy.concatenate(send_buffer_times).astype("int"))
+                numpy.concatenate(send_buffer_times).astype(numpy.uint32))
             if len(counts):
                 return int(numpy.max(counts))
             return 0
